requirements.txt
numpy
csv
os
matplotlib.pyplot